requirements.txt
torch
torchvision
streamlit
numpy
opencv-python-headless
pillow
matplotlib
tensorboard
imageio[ffmpeg]
flake8
